requirements.txt
aiogram>=3.4.0
httpx>=0.27.0
python-dotenv>=1.0.0
aiosqlite>=0.19.0
pydantic>=2.7.0

